Removes extra-deps requirement
Realized that the reason for the extra whitespace was that some of the
results were being passed through `pShow` twice which was escaping some
characters and adding whitespace. This commit prevents that behavior by
replacing one usage of pShow with a composition of show and Text.pack
which prevents that from happening.