Adds support for URLEncoded request bodies on the requestData field
To set a URLEncoded body set bodyType to urlencoded:

  requestData:
     bodyType: urlencoded
     content:
       a: b
       c: d
       e: f

To set a JSON body set bodyType to json:

  requestData:
     bodyType: json
     content:
       a: b
       c: d
       e: f

In either case the body is formed from the value of the content key.

The change is backwards compatible so that the following:

  requestData:
    a: b
    c: d
    e: A

will send a JSON body formed from the value of the requestData key.